more lab3

+++ Lab3/src/TIMERA0INT.h
@@ -6,7 +6,7 @@
 void config_pwm_timer(void);
 void start_pwm(uint8_t duty_cycle);
 void stop_pwm(void);
-void timerA0_config(unsigned short period);
+void timerA0_config(float period);
 void timerA0_IRQHandler(void);
 volatile static int flag;
